requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
nltk
dask[dataframe]
transformers
datasets
torch
tensorflow
newspaper3k
beautifulsoup4
requests
kagglehub
os